Add a test for v1/results with the needed fixtures
Adds a module for performing REST requests against Minder.

Adds a module for working with projects that we use to create a project
as a fixture and then delete them afterwards.

Adds a module for working with providers which currently just finds the
first github-app provider.

Uses these modules to build a very basic test for the v1/results API
that currently just makes sure that we get a valid empty JSON that
contains an empty list keyed at `entities`.

Related: https://github.com/stacklok/minder-stories/issues/59