Add tests for `ListRepositories2`.
This change adds tests for `ListRepositories2`, behind
`/api/v1/repositories` endpoint. In order to test them, a `GitHub`
class was also added that wraps `gh` CLI tool so that repositories can
be copied as we do in the original smoke tests.

Additional per-test teardown is based on the second code block of this
issue https://github.com/robotframework/robotframework/issues/4098

Fixes stacklok/minder-smoke-tests#240